Do not lowercase words if given in a hashref
There was a bug in actually adding words to the object was applying
lc() to all words.  Since hashref words are read in as-is and assumed
to be pre-split, it was causing any uppercase hash words to not get
their counts.  I removed the lc() in that case rather than making all
hashref entries lowercased to match, to stick to what seemed the spirit
of the hashref documentation saying that everything given that way gets
no further processing.